cloudfunction/gke_healthcheck/src/requirements.txt
slackweb==1.0.5
kubernetes==21.7.0
google-cloud-container==2.10.1